refactor: 동아리 홈 섹션을 위젯으로 분리
